+++ CHANGELOG.md
@@ -6,7 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- GenerateACI accepts contract interface again
 ### Changed
+- GenerateACI returns ACI formatted the same way as in aesophia\_cli
 ### Removed
 
 ## [7.3.0]